- Вынос работы с загрузкой программ из центра управления. - Подготовка базиса для разденого зранения набора инструментов. - Упрощение процесса добаления загруженной проги в модель
