+++ BackEnd/databasereader.h
@@ -14,7 +14,9 @@ class DataBaseReader : public QObject
     Q_OBJECT
 public:
     DataBaseReader(const QString& pathToDb);
+
     void slotSendQuery(const QString& queryStr, int valueNumbersAwaited = 1);
+
     QList<QVariantList> slotSendSelectQuery(const QStringList &tables,
                              const QStringList& columns,
                              const QString& conditions);
@@ -27,11 +29,10 @@ public:
     bool executeUpdateQuery(const QString& queryStr);
 
 signals:
-    // void signalResultReady(const QString& query, const QVariantList res);
+
     void signalResultReady(const QString& query, const QList<QVariantList> res);
 
 private:
-    // QSharedPointer<QSqlDatabase> dbPointer;
 
 };
