+++ BackEnd/proghandle.cpp
@@ -100,3 +100,4 @@ void ProgHandle::setScopeNameList(QMap<int, QString> scopes)
     setScopeIdx(0);
     emit scopeNameList();
 }
+
